Move Outbound tests to correct namespace
